+++ src/gui/gui-buffer.h
@@ -407,7 +407,7 @@ extern struct t_gui_buffer *gui_buffer_search (const char *plugin, const char *n
 extern struct t_gui_buffer *gui_buffer_search_by_partial_name (const char *plugin,
                                                                const char *name);
 extern struct t_gui_buffer *gui_buffer_search_by_number (int number);
-extern struct t_gui_buffer *gui_buffer_search_by_number_or_name (const char *string);
+extern struct t_gui_buffer *gui_buffer_search_by_id_number_name (const char *string);
 extern int gui_buffer_count_merged_buffers (int number);
 extern void gui_buffer_clear (struct t_gui_buffer *buffer);
 extern void gui_buffer_clear_all ();